% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealions-dataset.R
\docType{data}
\name{sealions}
\alias{sealions}
\title{Sea lions dataset}
\format{
data.frame with 734 observations and 21 variables:
\itemize{
\item \code{isla}:...
\item \code{lobera}:...
\item \code{codigo}:...
\item \code{latitud}:...
\item \code{longitud}:...
\item \code{tipologia}:...
\item \code{importancia}:...
\item \code{year}:...
\item \code{macho}:...
\item \code{hembra}:...
\item \code{adultos}:...
\item \code{agua}:...
\item \code{juveniles}:...
\item \code{cachorros_tierra}:...
\item \code{cachorros_agua}:...
\item \code{desconocido}:...
\item \code{total}:...
\item \code{total_adultos}:...
\item \code{total_cachorros}:...
\item \code{autor}:...
\item \code{status}:...
}
}
\source{
Under construction
}
\usage{
sealions
}
\description{
Dataset from the Juan Fernández archipelago islands
}
\keyword{datasets}
