% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammals-package.R
\docType{package}
\name{mammals-package}
\alias{mammals}
\alias{mammals-package}
\title{Population Dynamics Analysis}
\description{
Tools for the implementation of population dynamics analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matcasti/mammals}
  \item Report bugs at \url{https://github.com/matcasti/mammals/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{m99castillo@gmail.com} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Renato Borras-Chavez \email{rborras@gmail.com} (\href{https://orcid.org/0000-0002-6415-2121}{ORCID})
  \item Carla Rivera Rebella \email{rodofitas@gmail.com} (\href{https://orcid.org/0000-0001-8163-3325}{ORCID})
}

}
\keyword{internal}
